step-function-integration/src/agents/hello-world/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.2.0",
    "strands-agents>=1.22.0",
    "boto3>=1.34.0",
]
